Move clanmember icon logic to clanmanager
