+++ runelite-client/src/main/resources/net/runelite/client/game/Captain_clan_rank.png
