+++ runelite-client/src/main/resources/net/runelite/client/game/Corporal_clan_rank.png
