+++ runelite-client/src/main/resources/net/runelite/client/game/Friend_clan_rank.png
