+++ runelite-client/src/main/resources/net/runelite/client/game/General_clan_rank.png
