+++ runelite-client/src/main/resources/net/runelite/client/game/Lieutenant_clan_rank.png
