+++ runelite-client/src/main/resources/net/runelite/client/game/Owner_clan_rank.png
