+++ runelite-client/src/main/resources/net/runelite/client/game/Recruit_clan_rank.png
