+++ runelite-client/src/main/resources/net/runelite/client/game/Sergeant_clan_rank.png
